Inline `apply_deposit_for_builder`
